teste_2/code/requirements.txt
netifaces==0.11.0
pika==1.3.2
psutil==5.9.5
python-decouple==3.8
